資安修改

+++ ISTI_CityNavigation/Manage/mHandler/DeleteMember.aspx.cs
@@ -14,6 +14,7 @@ namespace ISTI_CityNavigation.Manage.mHandler
         Member_DB m_db = new Member_DB();
         MemberLog_DB ml_db = new MemberLog_DB();
         string id;
+        Common com = new Common();
         protected void Page_Load(object sender, EventArgs e)
         {
             ///-----------------------------------------------------
@@ -40,15 +41,24 @@ namespace ISTI_CityNavigation.Manage.mHandler
                 #endregion
 
                 id = (string.IsNullOrEmpty(Request["id"])) ? "" : Request["id"].ToString().Trim();
+                string token = (string.IsNullOrEmpty(Request["Token"])) ? "" : Request["Token"].ToString().Trim();
+                if (com.VeriftyToken(token))
+                {
+                    m_db._M_ID = id;
+                    m_db._M_ModId = LogInfo.mGuid;
+                    m_db._M_ModName = LogInfo.name;
+                    m_db.DeleteMember();
+                    addLog();
 
-                m_db._M_ID = id;
-                m_db._M_ModId = LogInfo.mGuid;
-                m_db._M_ModName = LogInfo.name;
-                m_db.DeleteMember();
-                addLog();
-
-                string xmlstr = "<?xml version='1.0' encoding='utf-8'?><root><Response>人員已刪除</Response></root>";
-                xDoc.LoadXml(xmlstr);
+                    string xmlstr = "<?xml version='1.0' encoding='utf-8'?><root><Response>人員已刪除</Response></root>";
+                    xDoc.LoadXml(xmlstr);
+                }
+                else
+                {
+                    xDoc = ExceptionUtil.GetTokenErrorMassageDocument();
+                    Response.ContentType = System.Net.Mime.MediaTypeNames.Text.Xml;
+                    xDoc.Save(Response.Output);
+                }
             }
             catch (Exception ex)
             {
